requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0
redis==5.0.1

# Security & Encryption
cryptography==41.0.7
python-dotenv==1.0.0

# Utils
aiohttp==3.9.1
pydantic==2.5.2
pydantic-settings==2.1.0